Add files via upload
This code was used to do a dive into the use of protocols in Swift. Protocols are a way of letting our client know: "if you want to use me that's cool, but you need to implement/do this first..." We are thus able to steal desirable behavior or characteristics from existing code while being able to avoid the potential pitfalls of inheritance.